count event attendees
